requirements.txt
Flask
gunicorn
numpy
matplotlib
pandas
scipy
scikit-learn
tensorflow
qiskit
psycopg2-binary
SQLAlchemy
requests
python-dotenv
rich
